rename owned row
